Replaces SVG logos with optimized and compressed versions.
As discussed _at length_ in the commit history of //HedronVision, the SVG logos
produced by Inkscape provide several opportunities to be compressed. This change
introduces compressed, cleaned versions of those SVGs to the heartrate website.

As a review, the primary changes are:

  - Cubic Bézier curves for which all control points were collinear are replaced
    with simple DrawLine commands.
  - Extraneous groups and attributes added by Inkscape are culled.
  - Deprecated SVG elements are removed.
  - Presentation attributes are deduplicated in a <style> section.
  - For the non-animated SVG, paths are merged according to symmetry.
  - For the animated SVG, the animation approach changed. Instead of animating
    the display attribute for each frame through several `none`s and one
    `inline`, instead we now animate a <use> element's href attribute through
    several frames defined in a <defs> section.
  - The SVGs are recentered on (0, 0) with a 500x500 viewBox.

Let's compare file sizes in bytes (using `wc -c < filename`):

cube: 2086 --> 1323 (63%)
cubeAnimated: 1133193 --> 189441 (17%)

Because we're on the web, we really care more about the change in gzipped size:

cube: 897 --> 718 (80%)
cubeAnimated: 145167 --> 65750 (45%)

If we really needed, we could compress these even more by removing comments and
uglifying the markup. However, we've already made huge wins - why ruin the
beautiful SVGs for just a little more?